Move ErrorCodeDecider mappings to ErrorCode
